+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.2] - 2024.1.4
+
+### Added
+
+- Added `HoloKitARKitNativeProvider` script to control the ARSession in native environment.
+- Added `WorldOriginResetter` script to reset the ARSession's world origin.
+- Added Reset World Origin sample.
+- Added `ARBackgroundVideoEnhancementManager` script to enable ARKit's 4K and HDR options.
+- Added AR Background Video Enhancement sample.
+
 ## [0.4.1] - 2024.1.3
 
 ### Added